split files in codegen
