+++ codegen/doc.go
@@ -1,2 +1,2 @@
-// Package codgen helps autogenerate tables for CloudQuery source plugins from Go structs (of relevant SDKs)
+// Package codegen helps autogenerate tables for CloudQuery source plugins from Go structs (of relevant SDKs)
 package codegen
