+++ codegen/table_test.go
@@ -6,11 +6,9 @@ import (
 	"testing"
 	"time"
 
-	"github.com/bradleyjkemp/cupaloy/v2"
 	"github.com/cloudquery/plugin-sdk/schema"
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
-	"github.com/stretchr/testify/require"
 )
 
 type embeddedStruct struct {
@@ -180,47 +178,6 @@ func TestTableFromGoStruct(t *testing.T) {
 				t.Fatal(err)
 			}
 			fmt.Println(buf.String())
-		})
-	}
-}
-
-func TestGenerateTemplate(t *testing.T) {
-	type args struct {
-		table *TableDefinition
-	}
-
-	tests := []struct {
-		name string
-		args args
-	}{
-		{
-			name: "should add comma between relations",
-			args: args{
-				table: &TableDefinition{
-					Name:      "with relations",
-					Relations: []string{"relation1", "relation2"},
-				},
-			},
-		},
-		{
-			name: "should add ignore_in_tests to columns",
-			args: args{
-				table: &TableDefinition{
-					Name: "with relations",
-					Columns: []ColumnDefinition{
-						{Name: "ignore_in_tests", Type: schema.TypeString, IgnoreInTests: true},
-					},
-				},
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			buf := bytes.NewBufferString("")
-			err := tt.args.table.GenerateTemplate(buf)
-			require.NoError(t, err)
-			cupaloy.SnapshotT(t, buf.Bytes())
 		})
 	}
 }
