+++ codegen/template.go
@@ -1,0 +1,23 @@
+package codegen
+
+import (
+	"embed"
+	"fmt"
+	"io"
+	"text/template"
+)
+
+//go:embed templates/*.go.tpl
+var TemplatesFS embed.FS
+
+func (t *TableDefinition) GenerateTemplate(wr io.Writer) error {
+	tpl, err := template.New("table.go.tpl").ParseFS(TemplatesFS, "templates/*")
+	if err != nil {
+		return fmt.Errorf("failed to parse template: %w", err)
+	}
+
+	if err := tpl.Execute(wr, t); err != nil {
+		return fmt.Errorf("failed to execute template: %w", err)
+	}
+	return nil
+}
