+++ codegen/util.go
@@ -1,0 +1,73 @@
+package codegen
+
+import (
+	"fmt"
+	"reflect"
+	"strings"
+	"time"
+
+	"github.com/cloudquery/plugin-sdk/schema"
+	"github.com/iancoleman/strcase"
+)
+
+func sliceContains(arr []string, s string) bool {
+	for _, v := range arr {
+		if v == s {
+			return true
+		}
+	}
+	return false
+}
+
+func isFieldStruct(reflectType reflect.Type) bool {
+	return reflectType.Kind() == reflect.Struct || (reflectType.Kind() == reflect.Ptr && reflectType.Elem().Kind() == reflect.Struct)
+}
+
+func valueToSchemaType(v reflect.Type) (schema.ValueType, error) {
+	k := v.Kind()
+	switch k {
+	case reflect.String:
+		return schema.TypeString, nil
+	case reflect.Bool:
+		return schema.TypeBool, nil
+	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64,
+		reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
+		return schema.TypeInt, nil
+	case reflect.Float32, reflect.Float64:
+		return schema.TypeFloat, nil
+	case reflect.Map:
+		return schema.TypeJSON, nil
+	case reflect.Struct:
+		timeValue := time.Time{}
+		if v == reflect.TypeOf(timeValue) {
+			return schema.TypeTimestamp, nil
+		}
+		return schema.TypeJSON, nil
+	case reflect.Pointer:
+		return valueToSchemaType(v.Elem())
+	case reflect.Slice:
+		switch v.Elem().Kind() {
+		case reflect.String:
+			return schema.TypeStringArray, nil
+		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64,
+			reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
+			return schema.TypeIntArray, nil
+		default:
+			return schema.TypeJSON, nil
+		}
+	default:
+		return schema.TypeInvalid, fmt.Errorf("unsupported type: %s", k)
+	}
+}
+
+func DefaultTransformer(field reflect.StructField) string {
+	name := field.Name
+	if jsonTag := strings.Split(field.Tag.Get("json"), ",")[0]; len(jsonTag) > 0 {
+		// return empty string if the field is not related api response
+		if jsonTag == "-" {
+			return ""
+		}
+		name = jsonTag
+	}
+	return strcase.ToSnake(name)
+}
